dnevni backup, ioc simulator

+++ Project/SkyPulse/Include/adc.h
@@ -1,6 +1,5 @@
 #ifndef					ADC_H
 #define					ADC_H
-#include				"stm32f2xx.h"
 
 typedef struct	{
 unsigned short	T2,T3,V5,V12,V24,cooler,bottle,compressor,air,Ipump;
@@ -31,9 +30,9 @@ int			f1=(ft[1]*(t[0]-to)-ft[0]*(t[1]-to)) / (t[0]-t[1]);
 #define	_V12to16X					(int)(12.0/_UREF*_Rdiv(820.0,3300.0)*65535.0+0.5)			
 #define	_V24to16X					(int)(24.0/_UREF*_Rdiv(820.0,6800.0)*65535.0+0.5)			
 	
-#define	_16XtoV5(a)				(float)((float)a/65535.0*_UREF/_Rdiv(820.0,820.0))			
-#define	_16XtoV12(a)			(float)((float)a/65535.0*_UREF/_Rdiv(820.0,3300.0))			
-#define	_16XtoV24(a)			(float)((float)a/65535.0*_UREF/_Rdiv(820.0,6800.0))			
+#define	_16XtoV5(a)				(double)((double)a/65535.0*_UREF/_Rdiv(820.0,820.0))			
+#define	_16XtoV12(a)			(double)((double)a/65535.0*_UREF/_Rdiv(820.0,3300.0))			
+#define	_16XtoV24(a)			(double)((double)a/65535.0*_UREF/_Rdiv(820.0,6800.0))			
 
 const int Ttab[]={ 1000, 2500, 5000, 8000 };
 const	int Rtab[]={ (0xffff*_Rdiv(18813.0,5100.0)), (0xffff*_Rdiv(10000.0,5100.0)), (0xffff*_Rdiv(3894.6,5100.0)), (0xffff*_Rdiv(1462.6,5100.0))};
